Add a floating point primitive to both the concrete and the What4 backends.
